refactor(utils): reorganize some chunks
