--- core/src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './haptic';
-export * from './keyboard';
-export * from './transition';
