+++ core/src/utils/overlays-interface.ts
@@ -43,17 +43,3 @@ export interface HTMLIonOverlayElement extends HTMLStencilElement {
 }
 
 export type OverlaySelect = HTMLIonActionSheetElement | HTMLIonAlertElement | HTMLIonPopoverElement;
-
-// TODO: uncomment when TS 3.0 issues are fixed
-// Overlay checks
-/*
-export type Conforms<T extends Required<B>, B> = T;
-export type HTMLOverlaysElement =
-  Conforms<Required<HTMLIonModalElement>, ModalOptions> |
-  Conforms<Required<HTMLIonToastElement>, ToastOptions> |
-  Conforms<Required<HTMLIonActionSheetElement>, ActionSheetOptions> |
-  Conforms<Required<HTMLIonAlertElement>, AlertOptions> |
-  Conforms<Required<HTMLIonPopoverElement>, PopoverOptions> |
-  Conforms<Required<HTMLIonPickerElement>, PickerOptions> |
-  Conforms<Required<HTMLIonLoadingElement>, LoadingOptions>;
-*/
